Move locale folder
